requirements.txt
anthropic>=0.92.0
pywebview>=5.0
pyserial>=3.5
